Update API bindings

+++ src/test/bot/utils/location.rs
@@ -13,7 +13,7 @@ impl LocationConfig {
         let random_latitude = self.latitude_bottom_right + (y_len * y_precent);
         let random_longitude = self.longitude_bottom_right + (x_len * x_precent);
 
-        Location::new(random_latitude as f32, random_longitude as f32)
+        Location::new(random_latitude, random_longitude)
     }
 
     pub fn middle_of_area(&self) -> Location {
@@ -26,6 +26,6 @@ impl LocationConfig {
         let latitude = self.latitude_bottom_right + (y_len * y_precent);
         let longitude = self.longitude_bottom_right + (x_len * x_precent);
 
-        Location::new(latitude as f32, longitude as f32)
+        Location::new(latitude, longitude)
     }
 }
